refactor: change folders to clear code
