+++ server/src/action/action.controller.ts
@@ -1,6 +1,8 @@
 import { Controller, Get, Param } from '@nestjs/common';
 import { ActionService } from './action.service';
+import { ApiTags } from '@nestjs/swagger';
 
+@ApiTags('actions')
 @Controller(':token/')
 export class ActionController {
   constructor(private authService: ActionService) {}
