--- server/src/mailing/mailing.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { MailingService } from './mailing.service';
-
-@Controller('mailing')
-export class MailingController {
-  constructor(readonly mailingService: MailingService) {}
-}
