+++ server/src/reaction/mailing/mailing.module.ts
@@ -1,12 +1,10 @@
 import { Module } from '@nestjs/common';
-import { MailingController } from './mailing.controller';
 import { MailingService } from './mailing.service';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaModule } from '../prisma/prisma.module';
+import { PrismaModule } from '../../prisma/prisma.module';
 
 @Module({
   imports: [ConfigModule.forRoot(), PrismaModule],
-  controllers: [MailingController],
   providers: [MailingService],
 })
 export class MailingModule {}
