+++ server/src/reaction/reaction.controller.ts
@@ -1,6 +1,8 @@
 import { Controller, Get, Param } from '@nestjs/common';
 import { ReactionService } from './reaction.service';
+import { ApiTags } from '@nestjs/swagger';
 
+@ApiTags('reactions')
 @Controller(':token/')
 export class ReactionController {
   constructor(private authService: ReactionService) {}
