+++ server/src/services/discord/discord.module.ts
@@ -2,13 +2,12 @@ import { Global, Module } from '@nestjs/common';
 import { DiscordController } from './discord.controller';
 import { DiscordService } from './discord.service';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaModule } from '../prisma/prisma.module';
-import { DiscordStrategy } from '../strategy/discord.strategy';
+import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Global()
 @Module({
   controllers: [DiscordController],
-  providers: [DiscordService, DiscordStrategy],
+  providers: [DiscordService],
   imports: [ConfigModule.forRoot(), PrismaModule],
 })
 export class DiscordModule {}
