+++ server/src/services/google-auth/google-auth.controller.ts
@@ -1,11 +1,12 @@
 import { GoogleOAuthGuard } from './google-oauth.guard';
 import { Controller, Get, Request, Res, UseGuards } from '@nestjs/common';
 import { GoogleAuthService } from './google-auth.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../../prisma/prisma.service';
 import { ForbiddenException } from '@nestjs/common';
 import { ServiceType } from '@prisma/client';
-import { ApiOperation } from '@nestjs/swagger';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
 
+@ApiTags('services')
 @Controller('myauth')
 export class GoogleAuthController {
   constructor(
