+++ server/src/services/services.module.ts
@@ -3,10 +3,13 @@ import { ServicesService } from './services.service';
 import { ServicesController } from './services.controller';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { MeService } from 'src/me/me.service';
+import { GithubAuthModule } from './github-auth/github-auth.module';
+import { GoogleAuthModule } from './google-auth/google-auth.module';
+import { DiscordModule } from './discord/discord.module';
 
 @Global()
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, GithubAuthModule, GoogleAuthModule, DiscordModule],
   providers: [ServicesService, MeService],
   controllers: [ServicesController],
 })
